+++ components/accordion/index.ts
@@ -1,0 +1,5 @@
+import Accordion from "./accordion";
+
+export type { AccordionProps, AccordionItem } from "./accordion";
+
+export default Accordion;