+++ components/index.ts
@@ -18,3 +18,5 @@ export { default as Progress } from "./progress";
 export type { ProgressProps } from "./progress";
 export { default as Tabs } from "./tabs";
 export type { TabsProps, TabsOption } from "./tabs";
+export { default as Accordion } from "./accordion";
+export type { AccordionProps, AccordionItem } from "./accordion";